requirement.txt
attrs==18.2.0
autopep8==1.4.2
certifi==2018.10.15
chardet==3.0.4
Django==1.11
djangorestframework==3.8.2
geographiclib==1.49
geopy==1.17.0
idna==2.7
pathlib-mate==0.0.14
pycodestyle==2.4.0
pytz==2018.7
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.13
urllib3==1.24
uszipcode==0.2.2
